last commit

+++ app/src/main/java/com/zm/englishtraining/ui/screens/StartScreen.kt
@@ -7,7 +7,7 @@ import com.zm.englishtraining.start_ui.StartNavigation
 
 class StartScreen(navController: NavController) : Screen(navController), StartNavigation {
 
-    override fun navigateToMode(topicIds: IntArray) {
+    override fun navigateToMode(topicIds: LongArray) {
         val data = ModeFragment.makeArgs(topicIds)
         navigateWithData(R.id.action_start_to_mode, data)
     }
